+++ stempo-auth/src/main/java/com/stempo/config/XssConfig.java
@@ -2,7 +2,7 @@ package com.stempo.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.stempo.filter.XssEscapeServletFilter;
-import com.stempo.filter.XssSanitizer;
+import com.stempo.util.XssSanitizer;
 import com.stempo.util.HtmlCharacterEscapes;
 import jakarta.servlet.Filter;
 import lombok.RequiredArgsConstructor;
