+++ stempo-auth/src/main/java/com/stempo/filter/XssEscapeServletRequestWrapper.java
@@ -1,5 +1,6 @@
 package com.stempo.filter;
 
+import com.stempo.util.XssSanitizer;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletRequestWrapper;
