+++ stempo-auth/src/main/java/com/stempo/util/XssSanitizer.java
@@ -1,4 +1,4 @@
-package com.stempo.filter;
+package com.stempo.util;
 
 import org.owasp.html.PolicyFactory;
 import org.owasp.html.Sanitizers;
